+++ CHANGES.rst
@@ -87,6 +87,8 @@ Cubeviz
 - Move slice slider to the plugin tray and add capability for selecting by wavelength as well as
   through a tool in the spectrum viewer. [#1013]
 
+- Cubeviz can now load 3D ndarray. [#1040]
+
 Imviz
 ^^^^^
 
@@ -160,6 +162,8 @@ Cubeviz
 - Moment Map plugin no longer crashes when writing out to FITS file. [#1099]
 
 - Moment Maps result is no longer rotated w.r.t. original data. [#1104]
+
+- Cubeviz parser is refactored for better support across all supported data formats. [#1040]
 
 Imviz
 ^^^^^